--- src/main/java/GreetingCommand.java
@@ -1,8 +1,0 @@
-public class GreetingCommand extends Command {
-    private static final String GREETING = "Hello! I am Duke\n" + "What can I do for you?\n";
-
-    @Override
-    public String action() {
-        return GREETING;
-    }
-}
