+++ src/main/java/duke/Ui.java
@@ -1,7 +1,9 @@
+package duke;
+
 public class Ui {
     private static final String LINE = "____________________________________________________________";
     private static final String BORDER = LINE + "\n";
-    private static final String GREETING = "Hello! I am Duke\n" + "What can I do for you?\n";
+    private static final String GREETING = "Hello! I am duke.\n" + "What can I do for you?\n";
     private static final String INVALID = "☹ OOPS!!! ";
 
     public void printMessage(String message) {
