+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,7 @@
+package duke.command;
+
+import duke.exception.DukeException;
+
 public abstract class Command {
     public abstract String action() throws DukeException;
 }
