+++ src/main/java/duke/command/CommandType.java
@@ -1,3 +1,6 @@
+package duke.command;
+
+import duke.exception.DukeException;
 
 public enum CommandType {
     BYE("bye"),
