+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public class ExitCommand extends Command {
     private static final String GOODBYE = "Bye. Hope to see you again soon!" + "\n";
