+++ src/main/java/duke/exception/DukeException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 import java.lang.Exception;
 
 public class DukeException extends Exception{
