+++ src/main/java/duke/task/Event.java
@@ -1,11 +1,15 @@
+package duke.task;
+
+import duke.exception.DukeException;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
+public class Event extends Task {
     protected LocalDate date;
 
-    public Deadline(String description, String date) throws DukeException{
+    public Event(String description, String date) throws DukeException {
         super(description);
         try {
             this.date = LocalDate.parse(date);
@@ -16,14 +20,14 @@ public class Deadline extends Task{
 
     @Override
     public String toString() {
-        return ("[D]" + super.toString() + " (by: "
+        return ("[E]" + super.toString() + " (at: "
                 + this.date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")");
     }
 
     @Override
     public String toStorageString() {
         String doneDescriptionString = super.toStorageString();
-        return "D" + Task.STORAGE_DELIMITER
+        return "E" + Task.STORAGE_DELIMITER
                 + doneDescriptionString + Task.STORAGE_DELIMITER
                 + this.date;
     }
