+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I am Duke
+Hello! I am duke.
 What can I do for you?
 ____________________________________________________________
